+++ git-odb/src/store/linked/find.rs
@@ -6,7 +6,7 @@ use git_pack::find::Entry;
 
 use crate::{
     pack,
-    pack::{bundle::Location, Bundle},
+    pack::Bundle,
     store::{compound, linked},
 };
 
@@ -115,33 +115,5 @@ impl crate::Find for linked::Store {
 
     fn try_find<'a>(&self, id: impl AsRef<oid>, buffer: &'a mut Vec<u8>) -> Result<Option<Data<'a>>, Self::Error> {
         pack::Find::try_find(self, id, buffer).map(|t| t.map(|t| t.0))
-    }
-}
-
-impl crate::pack::Find for &linked::Store {
-    type Error = compound::find::Error;
-
-    fn contains(&self, id: impl AsRef<oid>) -> bool {
-        (*self).contains(id)
-    }
-
-    fn try_find<'a>(
-        &self,
-        id: impl AsRef<oid>,
-        buffer: &'a mut Vec<u8>,
-    ) -> Result<Option<(git_object::Data<'a>, Option<pack::bundle::Location>)>, Self::Error> {
-        (*self).try_find(id, buffer)
-    }
-
-    fn location_by_oid(&self, id: impl AsRef<oid>, buf: &mut Vec<u8>) -> Option<Location> {
-        (*self).location_by_oid(id, buf)
-    }
-
-    fn bundle_by_pack_id(&self, pack_id: u32) -> Option<&Bundle> {
-        (*self).bundle_by_pack_id(pack_id)
-    }
-
-    fn entry_by_location(&self, location: &pack::bundle::Location) -> Option<Entry<'_>> {
-        (*self).entry_by_location(location)
     }
 }
